+++ src/main/cpp/vm/models/g1Heap.cpp
@@ -7,7 +7,7 @@ using namespace std;
 void G1Heap::iterate(const std::shared_ptr<JVM> &jvm, const std::shared_ptr<JvmTypesContainer> &typesContainer) {
     G1HeapRegionTable *pTable = hrm->getRegions();
     bool *compressedPointers = (bool *) jvm->getJVMFlagValue("UseCompressedClassPointers");
-    for (int i = 0; i < pTable->getLength(); i++) {
+    for (size_t i = 0; i < pTable->getLength(); i++) {
         HeapRegion *pRegion = pTable->getHeapRegionByIndex(i, typesContainer);
         if (pRegion == nullptr) break;
         auto bottom = pRegion->getBottom();
